Start splitting the BagVerifier monolith into separate traits
