+++ bag_verifier/src/main/scala/uk/ac/wellcome/platform/archive/bagverifier/verify/steps/Step.scala
@@ -1,0 +1,10 @@
+package uk.ac.wellcome.platform.archive.bagverifier.verify.steps
+
+trait Step {
+  case class BagVerifierError(
+    e: Throwable,
+    userMessage: Option[String] = None
+  )
+
+  type InternalResult[T] = Either[BagVerifierError, T]
+}
